The collector app can now actually collect and write data
This is still not working, but now when it runs it can actually call the GitHub API and write some files locally.  Awesome!

However, the functionality is still lacking the evaluation of the tree for the parents so this isn't able to actually save data properly in the nesting.

The next step will be to build the proper tree and evaluate it in the correct order tracking the parent relationships appropriately so the application can save data properly.

This is still a big step, somethign that looks like it almost works.