A mostly working way to save out the files
This took a long time, isn't even clost to fully tested, but this was incredibly complicated to design and it is finally to the point where it works enough that it is a good stopping point.

This has the main execution algorithm for the collectors.  It requires that a tree structure be built and the collectors be evaluated iteratively so that the correct order can be executed.

The thing is the retrieval of data is still not correct for non parent references.  That algorithm still needs to be implemented.

However, this is a big milestone for the collectors, it can actually collect some basic data!