An enhanced method for loading in data
This required a significantly more complex algorithm to flexibly navigate the tree of files on the file system to properly identify those that needed to be loaded.  There are a variety of conditions depending on the relations between the files and the parent structure for the classes and this implementation is intended to be very flexible at handling those.

This still needs to be more throughly unit tested but this implementation should unblock some additional development in building functional collectors.

The one thing that does come to mind is the need for manifest files at each level to properly inventory the files that have been generated.  This would simplify the error handling of the FileDataManager at this stage as it needs to make assumptions in loading data if other files or folders are present.  However, for the successive stages (Analyze and Report) the need for having the dependency tree be analyzed given the required inputs from prior phases adds additional complexity and the data storage and retrieval is intended to be flexible for all stages.